requirements.txt
awsebcli==3.12.4
botocore==1.10.19
cement==2.8.2
colorama==0.3.7
Django==1.11.15
django-filter==1.1.0
djangorestframework==3.7.7
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
jmespath==0.9.3
pathspec==0.5.5
psycopg2==2.7.4
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
websocket-client==0.47.0
